+++ app/controllers/Application.scala
@@ -29,7 +29,7 @@ class Application extends Controller{
   }
 
   def postData = Action { implicit request =>
-    println("request is :" + request.toString)
+    Logger("request is :" + request.toString)
     dataForm.bindFromRequest.fold(
       formWithErrors => Ok("commit error"),
       {
@@ -38,16 +38,7 @@ class Application extends Controller{
     )
   }
   
-  def calWithSleep(time: Int) : Int = {
-    Thread sleep time
-    time
-  }
   
-  //asyn action,sleep for time microseconds,and then return the time
-  def asynAct(time: Int) = Action.async {request =>
-    val futureInt = scala.concurrent.Future { calWithSleep(time) }
-    futureInt.map(i => Ok(views.html.AsyncRet(i.toString)))
-  }
 
 }
